// src/common/component-lib/storage-manager/storage.types.ts
export enum StorageKey {
  USER_DETAILS = "USER_DETAILS",
  DASH_BOARD_DETAILS = "DASH_BOARD_DETAILS",
  USER_VIEWED_LANDING_PAGE = "USER_VIEWED_LANDING_PAGE",
  TASKS_DETAILS = "TASKS_DETAILS",
  ACTIVITY_LOG_DETAILS = "ACTIVITY_LOG_DETAILS",
  SETTING = "SETTING",
  SELECTED_LANGUAGE = "SELECTED_LANGUAGE",
}